Move organization resource
