+++ tests/utils/fixtures/mock_auth_factor_totp.py
@@ -1,6 +1,6 @@
 import datetime
 
-from workos.resources.mfa import AuthenticationFactorTotp, ExtendedTotpFactor
+from workos.types.mfa import AuthenticationFactorTotp, ExtendedTotpFactor
 
 
 class MockAuthenticationFactorTotp(AuthenticationFactorTotp):
